givensi...

+++ QRparallel_experimental.c
@@ -435,7 +435,7 @@ int main(int argc, char* argv[]){
 			if(J[k] == J[i] || cabs(G[i+M*k]) < EPSILON ) continue;
 
 			first_non_zero_idx = i;
-			if(i == k+1) break; 	// no swapping needed, everythinig already in the right position
+			if(i == k+1) break; 	// no swapping needed, everything already in the right position
 
 
 			// else, swap rows i <-> k+1
@@ -495,7 +495,7 @@ int main(int argc, char* argv[]){
 
 					int nthreads_loc = np/(2*offset);
 					if(nthreads_loc == 0) nthreads_loc = 1;
-					else if ( nthreads_loc > (int) ((omp_get_max_threads()-2)*rp) ) nthreads_loc = (int) (int) ((omp_get_max_threads()-2)*rp);
+					else if ( nthreads_loc > (int) ((omp_get_max_threads()-2)*rp) ) nthreads_loc = (int) ((omp_get_max_threads()-2)*rp);
 
 					#pragma omp parallel for num_threads( nthreads_loc )
 					for(i = 0; i < np - offset; i += 2*offset){
